# platform/api/openapi/v1/horecaos-api.storefront.json
{
  "components" : {
    "schemas" : {
...
      "AddressFields" : {
        "properties" : {
          "apartment" : {
            "maxLength" : 32,
            "minLength" : 0,
            "type" : "string"
          },
          "city" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "district" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "entrance" : {
            "maxLength" : 32,
            "minLength" : 0,
            "type" : "string"
          },
          "floor" : {
            "maxLength" : 32,
            "minLength" : 0,
            "type" : "string"
          },
          "landmark" : {
            "maxLength" : 300,
            "minLength" : 0,
            "type" : "string"
          },
          "line1" : {
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          },
          "line2" : {
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          },
          "postalCode" : {
            "maxLength" : 32,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "CallbackRequestResponse" : {
        "properties" : {
          "leadId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "required" : [ "leadId" ],
        "type" : "object"
      },
...
      "ContactPointSummary" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "primary" : {
            "type" : "boolean"
          },
          "type" : {
            "enum" : [ "PHONE", "EMAIL" ],
            "type" : "string"
          },
          "verificationStatus" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "CustomerCallbackRequest" : {
        "properties" : {
          "forCompany" : {
            "type" : "boolean"
          },
          "note" : {
            "maxLength" : 1000,
            "minLength" : 0,
            "type" : "string"
          },
          "phone" : {
            "maxLength" : 32,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "phone" ],
        "type" : "object"
      },
...
      "LocationProfile" : {
        "properties" : {
          "addressLine" : {
            "type" : "string"
          },
          "city" : {
            "type" : "string"
          },
          "contactPhone" : {
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "district" : {
            "type" : "string"
          },
          "landmark" : {
            "type" : "string"
          },
          "latitude" : {
            "format" : "double",
            "type" : "number"
          },
          "longitude" : {
            "format" : "double",
            "type" : "number"
          }
        },
        "type" : "object"
      },
...
      "ProfileResponse" : {
        "properties" : {
          "accountId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "brandId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "contactPoints" : {
            "items" : {
              "$ref" : "#/components/schemas/ContactPointSummary"
            },
            "type" : "array"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "identityMode" : {
            "type" : "string"
          },
          "identityPolicyVersion" : {
            "format" : "int32",
            "type" : "integer"
          },
          "preferredLocale" : {
            "type" : "string"
          },
          "preferredTimezone" : {
            "type" : "string"
          },
          "profileScope" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "version" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "SaveAddressRequest" : {
        "properties" : {
          "coordinateSource" : {
            "enum" : [ "NOT_GEOCODED", "LANDMARK_ONLY", "GEOCODER", "CUSTOMER_PIN", "OPERATOR_PIN", "LEGACY_UNSOURCED" ],
            "type" : "string"
          },
          "deliveryInstructions" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          },
          "fields" : {
            "$ref" : "#/components/schemas/AddressFields"
          },
          "label" : {
            "maxLength" : 64,
            "minLength" : 0,
            "type" : "string"
          },
          "latitude" : {
            "format" : "double",
            "maximum" : 90,
            "minimum" : -90,
            "type" : "number"
          },
          "longitude" : {
            "format" : "double",
            "maximum" : 180,
            "minimum" : -180,
            "type" : "number"
          }
        },
        "required" : [ "coordinateSource", "fields" ],
        "type" : "object"
      },
...
      "StorefrontCustomerControllerAddressResponse" : {
        "properties" : {
          "addressId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "coordinateSource" : {
            "enum" : [ "NOT_GEOCODED", "LANDMARK_ONLY", "GEOCODER", "CUSTOMER_PIN", "OPERATOR_PIN", "LEGACY_UNSOURCED" ],
            "type" : "string"
          },
          "deliveryInstructions" : {
            "type" : "string"
          },
          "fields" : {
            "$ref" : "#/components/schemas/AddressFields"
          },
          "label" : {
            "type" : "string"
          },
          "latitude" : {
            "format" : "double",
            "type" : "number"
          },
          "longitude" : {
            "format" : "double",
            "type" : "number"
          },
          "version" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "StorefrontCustomerControllerUpdateProfileRequest" : {
        "properties" : {
          "displayName" : {
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          },
          "preferredLocale" : {
            "maxLength" : 16,
            "minLength" : 0,
            "type" : "string"
          },
          "preferredTimezone" : {
            "maxLength" : 64,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Keycloak access token issued for the horecaos-api audience",
        "scheme" : "bearer",
        "type" : "http"
      },
      "storefrontAppId" : {
        "description" : "The registered app's id. An identifier, not a secret: it appears in a browser bundle by design. Sent alongside the customer's session, never instead of it.",
        "in" : "header",
        "name" : "X-Storefront-App-Id",
        "type" : "apiKey"
      },
      "storefrontAppSecret" : {
        "description" : "A confidential client's secret. Server-side only; a public client has none and must not send one.",
        "in" : "header",
        "name" : "X-Storefront-App-Secret",
        "type" : "apiKey"
      }
    }
  },
  "info" : {
...
    "title" : "HorecaOS Platform API",
    "version" : "v1",
    "x-horecaos-contract" : {
      "appIdHeader" : "X-Storefront-App-Id",
      "appSecretHeader" : "X-Storefront-App-Secret",
      "deprecationWindowMonths" : 12,
      "publishedPathsNeverRemovedWithin" : "v1"
    }
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/storefront/tenants/{tenantId}/brands/{brandId}/locations/{locationId}/profile" : {
      "get" : {
        "description" : "Only fields the tenant already publishes on this branch elsewhere -- the same name and address shown in pickup search and printed on a receipt. Not found for an unknown, inactive, or cross-tenant location id.",
        "operationId" : "profile_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "tenantId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "brandId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "locationId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LocationProfile"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "A branch's own published display name and address",
        "tags" : [ "Storefront location profile" ],
        "x-horecaos-surface" : "published"
      }
    },
...
    "/api/v1/storefront/tenants/{tenantId}/brands/{brandId}/me" : {
      "get" : {
        "description" : "Includes the account id, which is otherwise unlearnable from the storefront, and the scope the profile is shared at. Contact points are listed by type and verification state and never by value.",
        "operationId" : "profile",
        "parameters" : [ {
          "in" : "path",
          "name" : "tenantId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "brandId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "The caller's own account and profile",
        "tags" : [ "Customer self-service" ],
        "x-horecaos-surface" : "published"
      },
      "patch" : {
        "description" : "The three fields a customer owns. Status, identity partition, merge target and policy version are not among them: each decides something about the account, and none is the account holder's to set. Under TENANT_SHARED the change is visible at every brand of the tenant, which is what profileScope says.",
        "operationId" : "updateProfile",
        "parameters" : [ {
          "in" : "path",
          "name" : "tenantId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "brandId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StorefrontCustomerControllerUpdateProfileRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Change the caller's own display name, language, or timezone",
        "tags" : [ "Customer self-service" ],
        "x-horecaos-surface" : "published"
      }
    },
    "/api/v1/storefront/tenants/{tenantId}/brands/{brandId}/me/addresses" : {
      "get" : {
        "description" : "Decrypted, because the person they belong to is asking, and the reveal is recorded against a purpose that says so. Archived addresses are not listed.",
        "operationId" : "addresses",
        "parameters" : [ {
          "in" : "path",
          "name" : "tenantId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "brandId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/StorefrontCustomerControllerAddressResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "The caller's own saved addresses",
        "tags" : [ "Customer self-service" ],
        "x-horecaos-surface" : "published"
      },
      "post" : {
        "description" : "The coordinate source is required and must agree with the coordinates: an address awaiting geocoding and one that legitimately has no point — a mahalla house given by its ориентир — are different states, and only one is worth retrying. A customer may claim CUSTOMER_PIN, NOT_GEOCODED or LANDMARK_ONLY; the operator and geocoder sources are not theirs to assert.",
        "operationId" : "addAddress",
        "parameters" : [ {
          "in" : "path",
          "name" : "tenantId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "brandId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SaveAddressRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StorefrontCustomerControllerAddressResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Save a new address",
        "tags" : [ "Customer self-service" ],
        "x-horecaos-surface" : "published"
      }
    },
...
    "/api/v1/storefront/tenants/{tenantId}/brands/{brandId}/me/callback-requests" : {
      "post" : {
        "description" : "Creates one lead in the call centre's queue, linked to the caller's own account. 202: what was created is a work item for a person, not a call. The number to ring is in the body. A guest asking on behalf of a company says so with forCompany=true, which files it as a catering enquiry.",
        "operationId" : "request",
        "parameters" : [ {
          "in" : "path",
          "name" : "tenantId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "brandId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CustomerCallbackRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CallbackRequestResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Ask the restaurant to call me back",
        "tags" : [ "Customer self-service" ],
        "x-horecaos-surface" : "published"
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ],
...
}